Allow appointments and meetings to be moved
The initial check to ensure the selected item is of type
Outlook.mailItem has been removed. Each object (obj) in the thread must
be of type mailItem, AppointmentItem, or MeetingItem in order to be
moved whereas before only mailItem was allowed.

This fixes the issue where having an AppointmentItem or MeetingItem in a
thread would cause the macro to return with the message "No folders
found."