Enable moving emails to subfolders of inbox, etc.
Rather than checking if "Inbox", "Sent Items", or "Calendar", was in the
folder string, the macro now checks if that is the exact folder. This
allows for emails to be moved to subfolders of "Inbox", "Sent Items", or
"Calendar".

Debugging was also added. To view debug messages, open the immediate
window under the View menu or with Ctrl+G.